Refactor types(case classes) into a single file
The reasoning is that you often just use go to definition to find these datatypes
