+++ data-generation/src/main/scala/factories/LandyGraphFactory.scala
@@ -2,9 +2,9 @@ package factories
 
 import org.apache.spark.graphx.{Edge, Graph}
 import org.apache.spark.sql.SparkSession
-import thesis.{LandyEdgePayload, LandyVertexPayload}
-import thesis.SnapshotDeltaObject.LandyAttributeGraph
+import thesis.DataTypes.LandyAttributeGraph
 import thesis.SparkConfiguration.getSparkSession
+import thesis.{LandyEdgePayload, LandyVertexPayload}
 
 import java.time.Instant
 import scala.collection.immutable.HashMap
@@ -18,9 +18,11 @@ object LandyGraphFactory {
 
     Graph(vertices, edges)
   }
+
   val t1 = Instant.parse("2000-01-01T00:00:00.000Z")
   val t2 = Instant.parse("2001-01-01T00:00:00.000Z")
   val t3 = Instant.parse("2002-01-01T00:00:00.000Z")
+
   def getVertices(): Seq[(Long, LandyVertexPayload)] = {
 
     Seq(
@@ -30,7 +32,7 @@ object LandyGraphFactory {
           validFrom = t1,
           validTo = t2,
           attributes = HashMap(
-            "color"->"red",
+            "color" -> "red",
           )
         )
       ),
@@ -40,7 +42,7 @@ object LandyGraphFactory {
           validFrom = t2,
           validTo = t3,
           attributes = HashMap(
-            "color"->"orange",
+            "color" -> "orange",
           )
         )
       ),
@@ -50,7 +52,7 @@ object LandyGraphFactory {
           validFrom = t1,
           validTo = t3,
           attributes = HashMap(
-            "color"->"red",
+            "color" -> "red",
           )
         )
       )
@@ -65,7 +67,7 @@ object LandyGraphFactory {
           validFrom = t1,
           validTo = t2,
           attributes = HashMap(
-            "relation"->"brother"
+            "relation" -> "brother"
           )
         )
       ),
@@ -75,7 +77,7 @@ object LandyGraphFactory {
           validFrom = t2,
           validTo = t3,
           attributes = HashMap(
-            "relation"->"brother"
+            "relation" -> "brother"
           )
         )
       )
